refactor(webpack): move manifest plugin types
