+++ development/webpack/utils/plugins/ManifestPlugin/types.ts
@@ -1,3 +1,4 @@
+import type { EntryOptions } from 'webpack';
 import type { Browser, Manifest } from '../../helpers';
 
 export type BundleSizeCategory =
@@ -22,6 +23,11 @@ export type BundleSizeStatsOptions = {
    */
   classifyEntrypoint: (name: string) => BundleSizeCategory | null;
 };
+
+export type EntryDescriptionNormalized = { import?: string[] } & Omit<
+  EntryOptions,
+  'name'
+>;
 
 export type BaseManifestPluginOptions<Zip extends boolean> = {
   /**
